utils/uploadscript/requirements.txt
requests
pyjwt
pytz